fix: change node configuration around the localizer
